Move plan Copy methods to their own file for easier diffing
